FEATURE: Move Session*Component to Middleware
