+++ Neos.Flow/Configuration/Settings.Http.yaml
@@ -29,6 +29,9 @@ Neos:
         'trustedProxies':
           position: 'start'
           middleware: 'Neos\Flow\Http\Middleware\TrustedProxiesMiddleware'
+        'session':
+          position: '10'
+          middleware: 'Neos\Flow\Http\Middleware\SessionMiddleware'
         'dispatch':
           position: 'end'
           middleware: 'Neos\Flow\Http\Middleware\ComponentChainMiddleware'
@@ -42,9 +45,6 @@ Neos:
         'preprocess':
           position: 'before process'
           chain:
-            'getSessionCookieFromRequest':
-              position: 'after trustedProxies'
-              component: 'Neos\Flow\Session\Http\SessionRequestComponent'
         'process':
           chain:
             'routing':
@@ -70,9 +70,6 @@ Neos:
 
         'postprocess':
           chain:
-            'setSessionCookie':
-              position: 'start'
-              component: 'Neos\Flow\Session\Http\SessionResponseComponent'
             'flashMessages':
               position: 'before setSessionCookie'
               component: 'Neos\Flow\Mvc\FlashMessage\FlashMessageComponent'
